[NA] minor fixes for guardrails (#1925)
* change GuardrailBatchItem to JSON views

* rename GuardrailBatchItem to Guardrail

* rename addGuardrailsBatch to createGuardrails

* removed unnecessary validation

* changed doubles to floats

* spotless

* fix compilation error

+++ apps/opik-backend/src/main/java/com/comet/opik/api/GuardrailTopicDetails.java
@@ -10,5 +10,5 @@ import java.util.Map;
 @Builder(toBuilder = true)
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record GuardrailTopicDetails(Map<String, Double> scores) {
+public record GuardrailTopicDetails(Map<String, Float> scores) {
 }
